+++ sickrage/core/webserver/handlers/api/v2/schedule/__init__.py
@@ -23,10 +23,10 @@ import datetime
 import sickrage
 from sickrage.core.helpers import convert_dict_keys_to_camelcase
 from sickrage.core.tv.show.coming_episodes import ComingEpisodes
-from sickrage.core.webserver.handlers.api import APIBaseHandler
+from sickrage.core.webserver.handlers.api.v2 import ApiV2BaseHandler
 
 
-class ApiV2ScheduleHandler(APIBaseHandler):
+class ApiV2ScheduleHandler(ApiV2BaseHandler):
     def get(self):
         """Get TV show schedule information"
         ---
@@ -74,4 +74,4 @@ class ApiV2ScheduleHandler(APIBaseHandler):
             results[i]['localtime'] = result['localtime'].timestamp()
             results[i] = convert_dict_keys_to_camelcase(results[i])
 
-        return self.write_json({'episodes': results, 'today': today.timestamp(), 'nextWeek': next_week.timestamp()})
+        return self.to_json({'episodes': results, 'today': today.timestamp(), 'nextWeek': next_week.timestamp()})