+++ sickrage/core/webserver/handlers/root.py
@@ -25,7 +25,6 @@ import os
 from functools import cmp_to_key
 
 from tornado.httputil import url_concat
-from tornado.ioloop import IOLoop
 from tornado.web import authenticated
 
 import sickrage
@@ -35,7 +34,7 @@ from sickrage.core.helpers import remove_article
 from sickrage.core.media.util import series_image, SeriesImageType
 from sickrage.core.tv.show.coming_episodes import ComingEpisodes, ComingEpsLayout, ComingEpsSortBy
 from sickrage.core.tv.show.helpers import get_show_list, find_show
-from sickrage.core.webserver import ApiHandler
+from sickrage.core.webserver.handlers.api.v1 import ApiV1Handler
 from sickrage.core.webserver.handlers.base import BaseHandler
 
 
@@ -45,7 +44,7 @@ class RobotsDotTxtHandler(BaseHandler):
 
     def get(self, *args, **kwargs):
         """ Keep web crawlers out """
-        return self.write("User-agent: *\nDisallow: /")
+        return "User-agent: *\nDisallow: /"
 
 
 class MessagesDotPoHandler(BaseHandler):
@@ -59,7 +58,7 @@ class MessagesDotPoHandler(BaseHandler):
             locale_file = os.path.join(sickrage.LOCALE_DIR, sickrage.app.config.gui.gui_lang, 'LC_MESSAGES/messages.po')
             if os.path.isfile(locale_file):
                 with open(locale_file, 'r', encoding='utf8') as f:
-                    return self.write(f.read())
+                    return f.read()
 
 class APIBulderHandler(BaseHandler):
     @authenticated
@@ -85,7 +84,7 @@ class APIBulderHandler(BaseHandler):
             apikey = _('API Key not generated')
 
         api_commands = {}
-        for command, api_call in ApiHandler(self.application, self.request).api_calls.items():
+        for command, api_call in ApiV1Handler(self.application, self.request).api_calls.items():
             api_commands[command] = api_call(self.application, self.request, **{'help': 1}).run()
 
         return self.render('api_builder.mako',
@@ -267,7 +266,7 @@ class QuicksearchDotJsonHandler(BaseHandler):
                 'seasons': 0,
             }]
 
-        return self.write(json.dumps(shows + episodes))
+        return json.dumps(shows + episodes)
 
 
 class ForceSchedulerJobHandler(BaseHandler):